Refactored Registry
No longer uses Entity objects but rather plain EIDs to improve the performance of the core ECS code.
Moved Component Containers out of Registry to improve readability.
Moved Entity handles to be managed by the World since they are the outward facing objects and don't need to exist in the core of the ECS.